refactor: Refactor code structure for improved readability and maintainability

+++ packages/create-fastymini-app/src/helper/applyPackageTemplate.ts
@@ -1,14 +1,10 @@
 import fs from 'fs-extra'
 import path from 'node:path'
-import { DatabaseType } from '../types'
-import { applyMongoosePlugin } from '../plugins/mongoose'
-
 interface ApplyPackageOptions {
   root: string
   projectName: string
   eslint: boolean
   prettier: boolean
-  database: DatabaseType
 }
 
 export async function applyPackageTemplate({
@@ -16,7 +12,7 @@ export async function applyPackageTemplate({
   projectName,
   eslint,
   prettier,
-  database,
+
 }: ApplyPackageOptions) {
   const tplPkg = path.join(root, 'package-template.json')
   const outPkg = path.join(root, 'package.json')
@@ -34,12 +30,9 @@ export async function applyPackageTemplate({
   const prettierIgnoreTpl = path.join(root, '.prettierignore-template')
 
   if (prettier) {
-    if (await fs.pathExists(prettierConfigTpl)) {
-      await fs.move(prettierConfigTpl, path.join(root, '.prettierrc'), { overwrite: true })
-    }
-    if (await fs.pathExists(prettierIgnoreTpl)) {
-      await fs.move(prettierIgnoreTpl, path.join(root, '.prettierignore'), { overwrite: true })
-    }
+    if (await fs.pathExists(prettierConfigTpl)) await fs.move(prettierConfigTpl, path.join(root, '.prettierrc'), { overwrite: true })
+    if (await fs.pathExists(prettierIgnoreTpl)) await fs.move(prettierIgnoreTpl, path.join(root, '.prettierignore'), { overwrite: true })
+    
   } else {
     await fs.remove(prettierConfigTpl)
     await fs.remove(prettierIgnoreTpl)
@@ -49,16 +42,9 @@ export async function applyPackageTemplate({
   const eslintOut = path.join(root, 'eslint.config.mjs')
 
   if (eslint) {
-    if (await fs.pathExists(eslintTpl)) {
-      await fs.move(eslintTpl, eslintOut, { overwrite: true })
-    }
+    if (await fs.pathExists(eslintTpl)) await fs.move(eslintTpl, eslintOut, { overwrite: true })
   } else {
     await fs.remove(eslintTpl)
     await fs.remove(eslintOut)
-  }
-
-  const useMongoose = Boolean(database === 'mongodb')
-  if (useMongoose) pkg.dependencies['mongoose'] = 'latest'
-  await applyMongoosePlugin({ root, enabled: useMongoose })
-
+  } 
 }
